Remove imputation sub-directory
