+++ src/pudl/analysis/__init__.py
@@ -10,7 +10,6 @@ from . import (
     allocate_gen_fuel,
     epacamd_eia,
     fuel_by_plant,
-    imputation,
     mcoe,
     plant_parts_eia,
     pudl_models,
