--- src/pudl/analysis/imputation/__init__.py
@@ -1,1 +1,0 @@
-"""Implement tooling to impute values in tables."""
