+++ test/unit/analysis/timeseries_cleaning_test.py
@@ -3,7 +3,7 @@
 import numpy as np
 import pytest
 
-import pudl.analysis.imputation.timeseries_cleaning
+import pudl.analysis.timeseries_cleaning
 
 
 def simulate_series(
@@ -90,7 +90,7 @@ def test_flags_and_imputes_anomalies(series_seed, anomalies_seed) -> None:
     values, indices = simulate_anomalies(x, seed=anomalies_seed)
     x.flat[indices] = values
     # Flag anomalies
-    s = pudl.analysis.imputation.timeseries_cleaning.Timeseries(x)
+    s = pudl.analysis.timeseries_cleaning.Timeseries(x)
     s.flag_ruggles()
     flag_indices = np.flatnonzero(~np.equal(s.flags, None))
     # Flag summary table has the right flag count
